Title: 두 수의 합, Time: 92ms, Memory: 8MB, Status: Wrong Answer - CodeTree

+++ 240627/두 수의 합/sum-of-two-num.java
@@ -1,27 +1,27 @@
 import java.io.*;
 import java.util.*;
 
-public class Main  {
+public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        HashMap<Integer, Integer> map = new HashMap<>();
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
-        int[] arr = new int[a];
-        int r = 0;
+        HashMap<Long, Long> map = new HashMap<>();
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
+        long[] arr = new long[(int) a];
+        long r = 0;
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < a; i++) {
-            int num = Integer.parseInt(st.nextToken());
-            map.put(num,b-num);
+            long num = Long.parseLong(st.nextToken());
+            map.put(num, b - num);
             arr[i] = num;
         }
         for (int i = 0; i < arr.length; i++) {
-            int num = map.get(arr[i]);
+            long num = map.get(arr[i]);
             if (map.containsKey(num)) {
                 r++;
             }
         }
-        System.out.println(r/2);
+        System.out.println(r / 2);
     }
 }
